+++ samples/DotNetCoreSample/Program.cs
@@ -16,7 +16,13 @@ namespace DotNetCoreSample
             FluentSettingsForExcel();
             var tempDirPath = $@"{Environment.GetEnvironmentVariable("USERPROFILE")}\Desktop\temp\test";
 
-            //var tempExcelPath = Path.Combine(tempDirPath, "test.xlsx");
+            //FluentSettings.For<ppDto>()
+            //    .HasSheetSetting(sheet =>
+            //    {
+            //        sheet.CellFilter = cell => cell.ColumnIndex <= 10;
+            //    });
+            var tempExcelPath = Path.Combine(tempDirPath, "testdata.xlsx");
+            var t_list = ExcelHelper.ToEntityList<ppDto>(tempExcelPath);
             //var tempTable = ExcelHelper.ToDataTable(tempExcelPath);
 
             //using (var conn = new SqlConnection("server=.;uid=liweihan;pwd=Admin888;database=Reservation"))
@@ -143,7 +149,12 @@ namespace DotNetCoreSample
                 .HasDescription("WeihanLi.Npoi test")
                 .HasSubject("WeihanLi.Npoi test");
 
-            setting.HasSheetConfiguration(0, "SystemSettingsList", 1, true);
+            setting.HasSheetSetting(config =>
+            {
+                config.StartRowIndex = 1;
+                config.SheetName = "SystemSettingsList";
+                config.AutoColumnWidthEnabled = true;
+            });
 
             // setting.HasFilter(0, 1).HasFreezePane(0, 1, 2, 1);