+++ src/WeihanLi.Npoi/Settings/SheetSetting.cs
@@ -1,26 +1,23 @@
-﻿using WeihanLi.Extensions;
+﻿using NPOI.SS.UserModel;
+using System;
+using WeihanLi.Extensions;
 
 namespace WeihanLi.Npoi.Settings
 {
-    internal sealed class SheetSetting
+    /// <summary>
+    /// Excel Sheet Settings
+    /// </summary>
+    public sealed class SheetSetting
     {
         private int _startRowIndex = 1;
         private string _sheetName = "Sheet0";
-        private int _sheetIndex;
+        private Func<IRow, bool> _rowFilter;
+        private Func<ICell, bool> _cellFilter;
 
-        /// <summary>
-        /// SheetIndex
-        /// </summary>
-        public int SheetIndex
+        public SheetSetting()
         {
-            get => _sheetIndex;
-            set
-            {
-                if (value >= 0)
-                {
-                    _sheetIndex = value;
-                }
-            }
+            _rowFilter = row => true;
+            _cellFilter = cell => true;
         }
 
         /// <summary>
@@ -67,5 +64,23 @@ namespace WeihanLi.Npoi.Settings
         /// enable auto column width
         /// </summary>
         public bool AutoColumnWidthEnabled { get; set; }
+
+        /// <summary>
+        /// Cell Filter
+        /// </summary>
+        public Func<ICell, bool> CellFilter
+        {
+            get => _cellFilter;
+            set => _cellFilter = value ?? (cell => true);
+        }
+
+        /// <summary>
+        /// Row Filter
+        /// </summary>
+        public Func<IRow, bool> RowFilter
+        {
+            get => _rowFilter;
+            set => _rowFilter = value ?? (row => true);
+        }
     }
 }