+++ docs/ReleaseNotes.md
@@ -1,5 +1,11 @@
 # WeihanLi.Npoi Release Notes
 
+## [1.12.0](https://www.nuget.org/packages/WeihanLi.Npoi/1.12.0)
+
+- refact `ExcelSetting` and `SheetSetting`
+- add support for `RowFilter` and `CellFilter`(mainly for import)
+- add support for reading file when file opened by another process
+
 ## [1.11.0](https://www.nuget.org/packages/WeihanLi.Npoi/1.11.0)
 
 - add support for formula value import